Compare user IDs before requesting profile from API
This commit changes ProfileController's behaviour so that it'll
check whether a user is trying to access their own profile page before
it decides which API request to make: if the IDs match, then it'll
call getOwnProfile; otherwise it'll call getUser.

getOwnProfile doesn't take any arguments (and DSAPI doesn't accept
any either) so this doesn't introduce a way for data to leak.