Cache sites in browser
This commit uses angular-cache to maintain a cache of lists of divesites,
slipways, and compressors in the user's browser (currently in-memory).
Adding new sites and updating existing ones clears the cached list of
sites of that type. This isn't optimal, but will work for the time being
(we should move toward caching and removing site-by-site).