Cache follow/follower lists
Cache follow/follower lists for users when they're first accessed. In
this commit, the cache is only invalidated when the signed-in user
follows/unfollows another user. In a future enhancement, we can use
server-push to update the cache when another user follows/unfollows the
signed-in user, but that's a niche use-case, since we're not expecting
users to spend hours and hours signed in to the site without doing a
page refresh.